+++ R/image.R
@@ -173,8 +173,8 @@ shrink_images = function(
 #' @param file Path to the image file to be uploaded.
 #' @param key Client ID for Imgur. It can be set via either the global option
 #'   `xfun.upload_imgur.key` or the environment variable
-#'   `R_XFUN_UPLOAD_IMGUR_KEY` (see [xfun::env_option()]). If neither is set,
-#'   this uses a client ID registered by Yihui Xie.
+#'   `R_XFUN_UPLOAD_IMGUR_KEY` (see [env_option()]). If neither is set, this
+#'   uses a client ID registered by Yihui Xie.
 #' @param use_curl Whether to use the R package \pkg{curl} to upload the image.
 #'   If `FALSE`, the system command `curl` will be used.
 #' @param include_xml Whether to include the XML response in the returned value.
@@ -183,9 +183,10 @@ shrink_images = function(
 #'   from Imgur (it will be parsed by \pkg{xml2} if available). See Imgur API in
 #'   the references.
 #' @author Yihui Xie, adapted from the \pkg{imguR} package by Aaron Statham
-#' @note Please register your own Imgur application to get your client ID; you
-#'   can certainly use mine, but this ID is in the public domain so everyone has
-#'   access to all images associated to it.
+#' @note Please register your own Imgur application to get your client ID. You
+#'   can certainly use mine, but this ID is in the public domain and it might
+#'   reach Imgur's rate limit if too many people are using it in the same time
+#'   period or someone is trying to upload too many images at once.
 #' @references A demo: <https://yihui.org/knitr/demo/upload/>
 #' @export
 #' @examples \dontrun{
@@ -201,10 +202,14 @@ shrink_images = function(
 #' options(xfun.upload_imgur.key = 'your imgur key')
 #' }
 upload_imgur = function(
-  file, key = env_option('xfun.upload_imgur.key', '9f3460e67f308f6'),
-  use_curl = loadable('curl'), include_xml = FALSE
+  file, key = env_option('xfun.upload_imgur.key'), use_curl = loadable('curl'),
+  include_xml = FALSE
 ) {
-  if (!is.character(key)) stop('The Imgur API Key must be a character string!')
+  if (is.null(key) || key == '' || key == '9f3460e67f308f6') {
+    i = floor(as.numeric(Sys.time()) %% 3) + 1
+    key = c("qyv(sq y 't)ws(", 'q (t"vxs%t "wr(', "xpyvq%vp'rwx)yq")[i]
+    key = decrypt(key, '1987052436fedcba')
+  }
   api = 'https://api.imgur.com/3/image.xml'
   hdr = paste('Authorization: Client-ID', key)
   if (use_curl) {
@@ -230,7 +235,7 @@ upload_imgur = function(
   if (length(link) != 1) stop(
     'Failed to upload ', file, sprintf(' (reason: %s)', if (is.character(res)) {
       grep_sub('.*<error>([^<]+)</error>.*', '\\1', res)
-    } else res[[1]]$error[[1]])
+    } else res[[c('data', 'error', 'message')]][[1]])
   )
   if (include_xml) structure(link, XML = res) else link
 }