+++ R/record.R
@@ -375,14 +375,18 @@ format.xfun_record_results = function(
         if (is_md && cls == 'plot') {
           z = sprintf('![%s](<%s>)', alt, z)
         } else {
+          o = attr(z, 'opts'); a = o$attr
           z = gsub('^(\\s*\n)+|\n\\s*$', '', one_string(z))  # trim blank lines
-          if (cls != 'source') z = paste('#>', split_lines(z))
+          if (cls != 'source') z = paste0(o$comment %||% '#> ', split_lines(z))
           if (is_md) {
             cls_all = if (any(c('message', 'warning', 'error') %in% cls_all)) {
               c('plain', cls_all)
             } else if (cls == 'source') {
               replace(cls_all, cls_all == 'source', 'r')
-            } else setdiff(cls_all, 'output')
+            } else {
+              if (length(a)) a = c(sub('^[.]', '', a), 'plain')
+              setdiff(c(a, cls_all), 'output')
+            }
             z = fenced_block(z, sprintf('.%s', cls_all))
           }
         }
@@ -391,7 +395,7 @@ format.xfun_record_results = function(
       if (!is_md) z = gsub('\n*$', '\n', z)
       z
     }))
-    return(raw_string(res))
+    return(raw_string(res, lang = if (is_md) '.md'))
   }
   res = unlist(lapply(x, function(z) {
     cls = sub('^record_', '', class(z))
@@ -415,7 +419,7 @@ format.xfun_record_results = function(
     res = sub('$body$', one_string(res), read_utf8(template), fixed = TRUE)
     if (length(x) > 0) res = sub('$title$', make_title(x[[1]]), res, fixed = TRUE)
   }
-  raw_string(res)
+  raw_string(res, lang = '.html')
 }
 
 # generate a title from the first line of a character vector